Implement digging for bricks
Digging does not yet work for cubes. Some restructuring may need
to be done to the shovel/chunk taking code. The way parameters are
passed around is a bit clunky (parameter hell, as it is affectionately
dubbed). Also, player does not yet have dirt added to their inventory.
It's not all put together. But it works and I am tired and I need to
go to bed so I have energy to study for finals this weekend.