Divorce completely from Lilboarder's Trench Digging gamemode
I have finally decided that this add-on should stand on its own. It adds enough
new functionality for me to feel it deserves distinction. In this commit,
functionality from Lilboarder's gamemode that was missing counterparts in Trench
Digging Plus was added. This includes speed digging and infinite digging.
Likewise, all trench related variables were put under the namespace TRT and TDP.
I admit this commit is a bit heavy handed. Many changes it introduced could
have and should have been implemented as seperate commits for the sake of
clarity and best practice. However, it is done now, and I personally do not
feel like trying to tear apart these changes, as I feel it would be a waste of
my energies (note that doing so would likely require the use of temporary code).